iot_data_generator/requirements.txt
paho-mqtt
pydantic
pydantic-settings
mysql-connector-python
azure-eventhub